Init checkin for Zoho CRM Telegram Bot
Init checkin for Zoho CRM Telegram Bot.
Create a telegram bot. Copy the bot-token.
One will need to create a login in Catalyst by Zoho .
Then create a project and in the project create an Advanced IO and a BasicIO function respectively.
Then the rest of the code can be copied and the edits made as per instructions in the file.
One would also need to generate the client_id and client_secret and redirect url in https://api-console.zoho.com using the server-based mechanism.
After deploying in Catalyst, go to BotFather and set the domain.
Once done, go to Postman and run the setwebhook command as mentioned in the index.js to link the Telegram Bot to the Catalyst project.